framescope/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
opencv-python==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
scipy==1.13.0
torch==2.3.0
transformers==4.41.0
accelerate==0.30.0
groq==0.9.0
pydantic==2.7.0
python-dotenv==1.0.1
pydantic-settings==2.0.0
